Added ViewPager fragments in the MyTickets screen
