+++ app/src/main/java/com/niupiao/niupiao/fragments/events/ComingSoonFragment.java
@@ -1,7 +1,6 @@
 package com.niupiao.niupiao.fragments.events;
 
 import android.app.Activity;
-import android.content.Context;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.util.Log;
@@ -10,18 +9,14 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.ListView;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import com.android.volley.VolleyError;
-import com.android.volley.toolbox.NetworkImageView;
-import com.niupiao.niupiao.Constants;
 import com.niupiao.niupiao.R;
-import com.niupiao.niupiao.adapters.ParcelableArrayAdapter;
+import com.niupiao.niupiao.adapters.EventsAdapter;
 import com.niupiao.niupiao.fragments.ViewPagerFragment;
 import com.niupiao.niupiao.models.Event;
 import com.niupiao.niupiao.requesters.EventsRequester;
-import com.niupiao.niupiao.utils.ImageLoaderHelper;
 import com.niupiao.niupiao.utils.SharedPrefsUtils;
 
 import java.util.List;
@@ -81,46 +76,6 @@ public class ComingSoonFragment extends ViewPagerFragment implements EventsReque
         args.putInt(POSITION_KEY, position);
         fragment.setArguments(args);
         return fragment;
-    }
-
-    class EventsAdapter extends ParcelableArrayAdapter<Event> {
-
-        public EventsAdapter(Context context) {
-            super(context, R.layout.list_view_item_concert);
-        }
-
-        class ViewHolder {
-            NetworkImageView image;
-            TextView date;
-            TextView name;
-            TextView ticketInfo;
-        }
-
-        @Override
-        public View getView(int position, View convertView, ViewGroup parent) {
-            ViewHolder viewHolder;
-            if (convertView == null) {
-                convertView = LayoutInflater.from(getContext()).inflate(R.layout.list_view_item_concert, parent, false);
-                viewHolder = new ViewHolder();
-                viewHolder.image = (NetworkImageView) convertView.findViewById(R.id.image);
-                viewHolder.date = (TextView) convertView.findViewById(R.id.date);
-                viewHolder.name = (TextView) convertView.findViewById(R.id.name);
-                viewHolder.ticketInfo = (TextView) convertView.findViewById(R.id.ticket_info);
-                convertView.setTag(viewHolder);
-            } else {
-                viewHolder = (ViewHolder) convertView.getTag();
-            }
-
-            Event event = getItem(position);
-
-            NetworkImageView image = viewHolder.image;
-            image.setImageUrl(Constants.Url.fullUrl(event.getImagePath()), ImageLoaderHelper.getInstance().getImageLoader());
-
-            viewHolder.date.setText(event.getDate());
-            viewHolder.name.setText(event.getName());
-            viewHolder.ticketInfo.setText(String.format("%d out of %d tickets sold", event.getTicketsSold(), event.getTotalTickets()));
-            return convertView;
-        }
     }
